SVY-18986 on restore, if there is onSort defined, use that for sorting, instead of simply applying the sort state to the foundset

+++ aggrid/groupingtable/groupingtable.js
@@ -378,7 +378,7 @@ angular.module('aggridGroupingtable', ['webSocketModule', 'servoy']).directive('
 								sortColumns.push(getColumnIndex(sortModel[i].colId));
 								sortColumnDirections.push(sortModel[i].sort);
 							}
-
+							refreshDatasource();
 							var sortHandlerPromise = $scope.handlers.onSort(sortColumns, sortColumnDirections);
 							sortHandlerPromises.push(sortHandlerPromise);
 							sortHandlerPromise.then(
@@ -711,7 +711,7 @@ angular.module('aggridGroupingtable', ['webSocketModule', 'servoy']).directive('
 								removeAllFoundsetRef = true;
 								gridOptions.api.purgeServerSideCache();
 							}
-							if($scope.handlers.onSort && isRenderedAndSelectionReady) {
+							if($scope.handlers.onSort) {
 								onSortHandler();
 							}
 						},
@@ -3484,52 +3484,55 @@ angular.module('aggridGroupingtable', ['webSocketModule', 'servoy']).directive('
 								state.rootGroupSort = sortModel[0];
 							}
 
-
-							// if there is no user defined onSort, set the sort from from foundset
-							var currentGridSort = getFoundsetSortModel(gridOptions.api.getSortModel());
-							var foundsetSort = stripUnsortableColumns(foundset.getSortColumns());
-							var isSortChanged = foundsetRefManager.isRoot && sortString != foundsetSort
-							&& currentGridSort.sortString != foundsetSort;
-	
-							if(isSortChanged) {
-								$log.debug('CHANGE SORT REQUEST');
-								var isColumnSortable = false;
-								// check sort columns in both the reques and model, because it is disable in the grid, it will be only in the model
-								var sortColumns = sortModel.concat(getSortModel());
-								for(var i = 0; i < sortColumns.length; i++) {
-									var col = gridOptions.columnApi.getColumn(sortColumns[i].colId);
-									if(col && col.getColDef().sortable) {
-										isColumnSortable = true;
-										break;
+							if(!$scope.handlers.onSort) {
+								// if there is no user defined onSort, set the sort from from foundset
+								var currentGridSort = getFoundsetSortModel(gridOptions.api.getSortModel());
+								var foundsetSort = stripUnsortableColumns(foundset.getSortColumns());
+								var isSortChanged = foundsetRefManager.isRoot && sortString != foundsetSort
+								&& currentGridSort.sortString != foundsetSort;
+		
+								if(isSortChanged) {
+									$log.debug('CHANGE SORT REQUEST');
+									var isColumnSortable = false;
+									// check sort columns in both the reques and model, because it is disable in the grid, it will be only in the model
+									var sortColumns = sortModel.concat(getSortModel());
+									for(var i = 0; i < sortColumns.length; i++) {
+										var col = gridOptions.columnApi.getColumn(sortColumns[i].colId);
+										if(col && col.getColDef().sortable) {
+											isColumnSortable = true;
+											break;
+										}
 									}
-								}
-	
-								if(isColumnSortable) {
-									// send sort request if header is clicked; skip if is is not from UI (isRenderedAndSelectionReady == false) or if it from a sort handler or a group column sort
-									if(isSortModelApplied) {
-										foundsetSortModel = getFoundsetSortModel(sortModel)
-										sortPromise = foundsetRefManager.sort(foundsetSortModel.sortColumns);
-										sortPromise.then(function() {
-											getDataFromFoundset(foundsetRefManager);
-											// give time to the foundset change listener to know it was a client side requested sort
-											setTimeout(function() {
-												sortPromise = null;
-											}, 0);
-										}).catch(function(e) {
-											sortPromise = null
-										});
+		
+									if(isColumnSortable) {
+										// send sort request if header is clicked; skip if is is not from UI (isRenderedAndSelectionReady == false) or if it from a sort handler or a group column sort
+										if(isSortModelApplied) {
+											foundsetSortModel = getFoundsetSortModel(sortModel)
+											sortPromise = foundsetRefManager.sort(foundsetSortModel.sortColumns);
+											sortPromise.then(function() {
+												getDataFromFoundset(foundsetRefManager);
+												// give time to the foundset change listener to know it was a client side requested sort
+												setTimeout(function() {
+													sortPromise = null;
+												}, 0);
+											}).catch(function(e) {
+												sortPromise = null
+											});
+										}
+										// set the grid sorting if foundset sort changed from the grid initialization (like doing foundset sort on form's onShow)
+										else {
+											applySortModel(getSortModel());
+											gridOptions.api.purgeServerSideCache();
+										}
 									}
-									// set the grid sorting if foundset sort changed from the grid initialization (like doing foundset sort on form's onShow)
 									else {
-										applySortModel(getSortModel());
-										gridOptions.api.purgeServerSideCache();
+										getDataFromFoundset(foundsetRefManager);
 									}
 								}
 								else {
 									getDataFromFoundset(foundsetRefManager);
 								}
-							}
-							else {
+							} else {
 								getDataFromFoundset(foundsetRefManager);
 							}
 						}
